+++ src/test/java/io/anserini/index/IndexReaderUtilsTest.java
@@ -386,25 +386,25 @@ public class IndexReaderUtilsTest extends IndexerTestBase {
     Directory dir = FSDirectory.open(tempDir1);
     IndexReader reader = DirectoryReader.open(dir);
 
-    Map<String, List<Long>> termPositions;
+    Map<String, List<Integer>> termPositions;
 
     termPositions = IndexReaderUtils.getTermPositions(reader, "doc1");
-    assertEquals(Long.valueOf(0), termPositions.get("here").get(0));
-    assertEquals(Long.valueOf(4), termPositions.get("here").get(1));
-    assertEquals(Long.valueOf(2), termPositions.get("some").get(0));
-    assertEquals(Long.valueOf(6), termPositions.get("some").get(1));
-    assertEquals(Long.valueOf(3), termPositions.get("text").get(0));
-    assertEquals(Long.valueOf(8), termPositions.get("text").get(1));
-    assertEquals(Long.valueOf(7), termPositions.get("more").get(0));
-    assertEquals(Long.valueOf(9), termPositions.get("citi").get(0));
+    assertEquals(Integer.valueOf(0), termPositions.get("here").get(0));
+    assertEquals(Integer.valueOf(4), termPositions.get("here").get(1));
+    assertEquals(Integer.valueOf(2), termPositions.get("some").get(0));
+    assertEquals(Integer.valueOf(6), termPositions.get("some").get(1));
+    assertEquals(Integer.valueOf(3), termPositions.get("text").get(0));
+    assertEquals(Integer.valueOf(8), termPositions.get("text").get(1));
+    assertEquals(Integer.valueOf(7), termPositions.get("more").get(0));
+    assertEquals(Integer.valueOf(9), termPositions.get("citi").get(0));
 
     termPositions = IndexReaderUtils.getTermPositions(reader, "doc2");
-    assertEquals(Long.valueOf(0), termPositions.get("more").get(0));
-    assertEquals(Long.valueOf(1), termPositions.get("text").get(0));
+    assertEquals(Integer.valueOf(0), termPositions.get("more").get(0));
+    assertEquals(Integer.valueOf(1), termPositions.get("text").get(0));
 
     termPositions = IndexReaderUtils.getTermPositions(reader, "doc3");
-    assertEquals(Long.valueOf(0), termPositions.get("here").get(0));
-    assertEquals(Long.valueOf(3), termPositions.get("test").get(0));
+    assertEquals(Integer.valueOf(0), termPositions.get("here").get(0));
+    assertEquals(Integer.valueOf(3), termPositions.get("test").get(0));
 
     // Invalid docid.
     assertTrue(IndexReaderUtils.getDocumentVector(reader, "foo") == null);
